refactor: move multiple streams stream to fluid domain
