+++ src/libecalc/domain/process/compressor/dto/train.py
@@ -2,7 +2,6 @@ from typing import Literal
 
 from libecalc.common.energy_model_type import EnergyModelType
 from libecalc.common.fixed_speed_pressure_control import FixedSpeedPressureControl
-from libecalc.common.multiple_streams_stream import MultipleStreamsAndPressureStream
 from libecalc.common.serializable_chart import SingleSpeedChartDTO, VariableSpeedChartDTO
 from libecalc.domain.component_validation_error import (
     ModelValidationError,
@@ -13,6 +12,7 @@ from libecalc.domain.component_validation_error import (
 from libecalc.domain.process.compressor.dto.stage import CompressorStage
 from libecalc.domain.process.dto.base import EnergyModel
 from libecalc.domain.process.value_objects.fluid_stream.fluid_model import FluidModel
+from libecalc.domain.process.value_objects.fluid_stream.multiple_streams_stream import MultipleStreamsAndPressureStream
 from libecalc.presentation.yaml.validation_errors import Location
