+++ tests/libecalc/core/models/compressor_modelling/conftest.py
@@ -6,7 +6,6 @@ import pytest
 
 import libecalc.common.serializable_chart
 from libecalc.common.fixed_speed_pressure_control import FixedSpeedPressureControl
-from libecalc.common.multiple_streams_stream import FluidStreamType, MultipleStreamsAndPressureStream
 from libecalc.common.units import Unit
 from libecalc.domain.process.compressor import dto
 from libecalc.domain.process.compressor.core.train.single_speed_compressor_train_common_shaft import (
@@ -22,6 +21,10 @@ from libecalc.domain.process.compressor.core.train.variable_speed_compressor_tra
 )
 from libecalc.domain.process.value_objects.chart.compressor import SingleSpeedCompressorChart
 from libecalc.domain.process.value_objects.fluid_stream.fluid_model import EoSModel, FluidComposition, FluidModel
+from libecalc.domain.process.value_objects.fluid_stream.multiple_streams_stream import (
+    FluidStreamType,
+    MultipleStreamsAndPressureStream,
+)
 from libecalc.infrastructure.neqsim_fluid_provider.neqsim_fluid_factory import NeqSimFluidFactory
